Lab-1: Add a simple version of the UNIX xargs
Write a simple version of the UNIX xargs program for xv6: its arguments
describe a command to run, it reads lines from the standard input, and
it runs the command for each line, appending the line to the command's
arguments. Your solution should be in the file user/xargs.c.